Use one editor, multiple models
