+++ src/components/CodeEditor.css
@@ -1,7 +1,7 @@
-.code-editor {
+/* .code-editor {
   width: 100%;
   height: 100%;
-}
+} */
 
 .editor-failed {
   width: 100%;
