+++ src/components/Tiles/Editor.css
@@ -1,4 +1,15 @@
 .editor {
   height: 100%;
-  z-index: -1;
+  position: relative;
+  overflow: hidden;
+}
+.code-editor {
+  width: 100%;
+  height: 100%;
+  max-height: 100%;
+  overflow: hidden;
+  position: relative;
+  position: fixed;
+  top: 4.45rem;
+  /* transform: translate(0, -100%); */
 }
