+++ src/components/Tiles/Editor.tsx
@@ -8,9 +8,62 @@ import {
   TabList,
   TabPanel,
   Tabs,
+  useColorScheme,
 } from "@mui/joy";
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 import CloseIcon from "@mui/icons-material/Close";
+import * as monaco from "monaco-editor";
+
+import { initialize } from "@codingame/monaco-vscode-api";
+import getLanguagesServiceOverride from "@codingame/monaco-vscode-languages-service-override";
+import getThemeServiceOverride from "@codingame/monaco-vscode-theme-service-override";
+import getTextMateServiceOverride from "@codingame/monaco-vscode-textmate-service-override";
+import getEditorServiceOverride from "@codingame/monaco-vscode-editor-service-override";
+import getModelServiceOverride from "@codingame/monaco-vscode-model-service-override";
+import "@codingame/monaco-vscode-swift-default-extension";
+import "@codingame/monaco-vscode-theme-defaults-default-extension";
+import "vscode/localExtensionHost";
+
+// adding worker
+export type WorkerLoader = () => Worker;
+const workerLoaders: Partial<Record<string, WorkerLoader>> = {
+  TextEditorWorker: () =>
+    new Worker(
+      new URL("monaco-editor/esm/vs/editor/editor.worker.js", import.meta.url),
+      { type: "module" }
+    ),
+  TextMateWorker: () =>
+    new Worker(
+      new URL(
+        "@codingame/monaco-vscode-textmate-service-override/worker",
+        import.meta.url
+      ),
+      { type: "module" }
+    ),
+};
+
+window.MonacoEnvironment = {
+  getWorker: function (_workerId, label) {
+    const workerFactory = workerLoaders[label];
+    if (workerFactory != null) {
+      return workerFactory();
+    }
+    throw new Error(`Worker ${label} not found`);
+  },
+};
+
+await initialize({
+  ...getTextMateServiceOverride(),
+  ...getThemeServiceOverride(),
+  ...getLanguagesServiceOverride(),
+  ...getEditorServiceOverride((a, b, c) => {
+    return new Promise((resolve) => {
+      console.log(a, b, c);
+      resolve(undefined);
+    });
+  }),
+  ...getModelServiceOverride(),
+});
 
 export interface EditorProps {
   openFiles: string[];
@@ -34,6 +87,12 @@ export default ({
   const [unsavedFiles, setUnsavedFiles] = useState<string[]>([]);
   const [focused, setFocused] = useState<number>();
   const editors = useRef<(CodeEditorHandles | null)[]>([]);
+  const [editor, setEditor] =
+    useState<monaco.editor.IStandaloneCodeEditor | null>(null);
+
+  const { mode } = useColorScheme();
+
+  const monacoEl = useRef(null);
 
   useEffect(() => {
     editors.current = editors.current.slice(0, openFiles.length);
@@ -68,6 +127,43 @@ export default ({
 
     fetchTabNames();
   }, [openFiles]);
+
+  useEffect(() => {
+    if (monacoEl.current && !editor) {
+      let colorScheme = mode;
+      if (colorScheme === "system") {
+        colorScheme = window.matchMedia("(prefers-color-scheme: dark)").matches
+          ? "dark"
+          : "light";
+      }
+
+      const newEditor = monaco.editor.create(monacoEl.current, {
+        value: "",
+        language: "plaintext",
+        theme: "vs-" + colorScheme,
+      });
+
+      setEditor(newEditor);
+
+      return () => {
+        newEditor.dispose();
+      };
+    }
+  }, []);
+
+  useEffect(() => {
+    if (!monacoEl.current || !editor) return;
+
+    const resizeObserver = new ResizeObserver(() => {
+      editor.layout();
+    });
+
+    resizeObserver.observe(monacoEl.current);
+
+    return () => {
+      resizeObserver.disconnect();
+    };
+  }, [editor]);
 
   return (
     <div className={"editor"}>
@@ -113,24 +209,31 @@ export default ({
           <TabPanel
             value={index}
             key={tab.file}
-            sx={{ padding: 0 }}
-            keepMounted
+            sx={{ padding: 0, width: 0, height: 0 }}
           >
-            <CodeEditor
-              ref={(el) => (editors.current[index] = el)}
-              key={tab.file}
-              file={tab.file}
-              setUnsaved={(unsaved: boolean) => {
-                if (unsaved)
-                  setUnsavedFiles((unsaved) => [...unsaved, tab.file]);
-                else
-                  setUnsavedFiles((unsaved) =>
-                    unsaved.filter((unsavedFile) => unsavedFile !== tab.file)
-                  );
-              }}
-            />
+            {editor && (
+              <CodeEditor
+                editor={editor}
+                ref={(el) => (editors.current[index] = el)}
+                key={tab.file}
+                file={tab.file}
+                setUnsaved={(unsaved: boolean) => {
+                  if (unsaved)
+                    setUnsavedFiles((unsaved) => [...unsaved, tab.file]);
+                  else
+                    setUnsavedFiles((unsaved) =>
+                      unsaved.filter((unsavedFile) => unsavedFile !== tab.file)
+                    );
+                }}
+              />
+            )}
           </TabPanel>
         ))}
+        <div
+          className={"code-editor"}
+          ref={monacoEl}
+          style={tabs.length >= 1 ? {} : { display: "none" }}
+        />
       </Tabs>
     </div>
   );
